+++ src/helpers/__tests__/path.ts
@@ -1,0 +1,5 @@
+import {url} from 'helpers/path';
+
+test('url works', () => {
+  expect(url('http://localhost', 'api', 'sample')).toEqual('http://localhost/api/sample');
+});