+++ src/helpers/path.ts
@@ -1,0 +1,5 @@
+import path from 'path';
+
+export function url(baseURL: string, ...urls: string[]) {
+  return `${baseURL}/${path.join(...urls)}`;
+}