+++ src/services/index.ts
@@ -1,3 +1,3 @@
-export * from './common-state';
+export * from 'services/common-service';
 
 export * from './translation-service';